+++ Adj/Program/Controlflow/Implementation/Optics.hs
@@ -1,9 +1,11 @@
 module Adj.Program.Controlflow.Implementation.Optics where
 
-import Adj.Auxiliary ((=-), type (=!?=), FG (FG))
+import Adj.Auxiliary ((=-), type (=!?=), FG_ (FG_))
 import Adj.Algebra.Category (type (<--), type (-->), type (:*:>), type (:+:>), (.), (..:), Opposite (Opposite), Straight (Straight), extract)
 import Adj.Algebra.Set ((:+:) (This, That), (:*:) ((:*:)))
 
+-- TODO: to use it as natural transformation
+-- I should hide source and target parameters
 type Lens queried required source target =
 	source --> (((:*:>) (queried target) =!?= (-->) (required target)) source)
 
@@ -19,8 +21,8 @@ type Prism available set subset =
 
 review :: Prism available set subset -> set -> available subset
 review prism set = case prism =- set of
-	FG (Straight (This subset)) -> subset
-	FG (Straight (That m)) -> m =- set
+	FG_ (Straight (This subset)) -> subset
+	FG_ (Straight (That m)) -> m =- set
 
 -- TODO: think about composition between different type of optics
 -- TODO: Lens and Prism types looks the same so maybe we can just generalize them somehow